+++ nix/packages/nxr.nix
@@ -8,7 +8,7 @@
 
 rustPlatform.buildRustPackage {
   pname = "nxr";
-  version = "3.5.2";
+  version = "3.5.3";
 
   inherit src;